requirements.txt
groq==0.5.0
huggingface-hub==0.22.2
notion-client==2.2.1
numpy==1.26.4
openai==1.23.6
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.33.0
yfinance==0.2.38
